# src/backend/api/openapi/swagger.yaml
basePath: /v1
definitions:
  handler.CreateRecipeRequest:
    properties:
      cookTime:
        type: string
      cuisineId:
        type: string
      description:
        type: string
      directions:
        items:
          type: string
        type: array
      ingredientIds:
        items:
          type: string
        type: array
      mainIngredientId:
        type: string
      name:
        type: string
      prepTime:
        type: string
    type: object
  handler.CuisineJSON:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  handler.DailyConstraint:
    properties:
      cuisineConstraints:
        items:
          $ref: '#/definitions/handler.EntityConstraint'
        type: array
      ingredientConstraints:
        items:
          $ref: '#/definitions/handler.EntityConstraint'
        type: array
    type: object
  handler.EntityConstraint:
    properties:
      entityId:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.IngredientJSON:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  handler.PaginatedRecipesJSON:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.RecipeJSON'
        type: array
      pageIndex:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
    type: object
  handler.RecipeJSON:
    properties:
      cookTime:
        type: string
      cuisine:
        $ref: '#/definitions/handler.CuisineJSON'
      description:
        type: string
      directions:
        items:
          type: string
        type: array
      id:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/handler.IngredientJSON'
        type: array
      mainIngredient:
        $ref: '#/definitions/handler.IngredientJSON'
      name:
        type: string
      prepTime:
        type: string
    type: object
  handler.SuggestRequest:
    properties:
      alreadySelectedRecipeIds:
        items:
          type: string
        type: array
      amount:
        type: integer
      dailyConstraints:
        items:
          $ref: '#/definitions/handler.DailyConstraint'
        type: array
    type: object
  handler.SuggestResponse:
    properties:
      recipeIds:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    name: PlatePilot Support
  description: Intelligent meal planning and recipe management API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PlatePilot API
  version: "1.0"
paths:
  /mealplan/suggest:
    post:
      consumes:
      - application/json
      description: Suggests recipes based on constraints and already selected recipes
      parameters:
      - description: Suggestion request with constraints
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SuggestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SuggestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Suggest recipes for meal planning
      tags:
      - mealplan
  /recipe/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a single recipe by its unique identifier
      parameters:
      - description: Recipe ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RecipeJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get recipe by ID
      tags:
      - recipes
  /recipe/all:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of all recipes
      parameters:
      - default: 1
        description: Page number (1-indexed)
        in: query
        name: pageIndex
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PaginatedRecipesJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get all recipes (paginated)
      tags:
      - recipes
  /recipe/allergy/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all recipes that do not contain a specific allergen
      parameters:
      - description: Allergy ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.RecipeJSON'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get recipes avoiding allergen
      tags:
      - recipes
  /recipe/create:
    post:
      consumes:
      - application/json
      description: Creates a new recipe with the provided details
      parameters:
      - description: Recipe to create
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/handler.CreateRecipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RecipeJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new recipe
      tags:
      - recipes
  /recipe/cuisine/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all recipes belonging to a specific cuisine
      parameters:
      - description: Cuisine ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.RecipeJSON'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get recipes by cuisine
      tags:
      - recipes
  /recipe/ingredient/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves all recipes containing a specific ingredient
      parameters:
      - description: Ingredient ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.RecipeJSON'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get recipes by ingredient
      tags:
      - recipes
  /recipe/similar:
    get:
      consumes:
      - application/json
      description: Finds recipes similar to the specified recipe using vector search
      parameters:
      - description: Recipe ID to find similar recipes for
        in: query
        name: recipe
        required: true
        type: string
      - default: 5
        description: Number of similar recipes to return (max 50)
        in: query
        name: amount
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.RecipeJSON'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get similar recipes
      tags:
      - recipes
schemes:
- http
- https
swagger: "2.0"